--- operator/pkg/compare/compare_test.go
@@ -1,969 +1,0 @@
-// Copyright Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package compare
-
-import (
-	"testing"
-)
-
-func TestYAMLCmp(t *testing.T) {
-	tests := []struct {
-		desc string
-		a    string
-		b    string
-		want any
-	}{
-		{
-			desc: "empty string into nil",
-			a:    `metadata: ""`,
-			b:    `metadata: `,
-			want: ``,
-		},
-		{
-			desc: "empty array into nil",
-			a:    `metadata: []`,
-			b:    `metadata: `,
-			want: ``,
-		},
-		{
-			desc: "empty map into nil",
-			a:    `metadata: {}`,
-			b:    `metadata: `,
-			want: ``,
-		},
-		{
-			desc: "two additional",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  namespace: istio-system
-  labels:
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			want: `metadata:
-  labels:
-    app: <empty> -> istio-ingressgateway (ADDED)
-  name: <empty> -> istio-ingressgateway (ADDED)
-`,
-		},
-		{
-			desc: "two missing",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  namespace: istio-system
-  labels:
-    release: istio`,
-			want: `metadata:
-  labels:
-    app: istio-ingressgateway -> <empty> (REMOVED)
-  name: istio-ingressgateway -> <empty> (REMOVED)
-`,
-		},
-		{
-			desc: "one missing",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    release: istio`,
-			want: `metadata:
-  labels:
-    app: istio-ingressgateway -> <empty> (REMOVED)
-`,
-		},
-		{
-			desc: "one additional",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			want: `metadata:
-  labels:
-    app: <empty> -> istio-ingressgateway (ADDED)
-`,
-		},
-		{
-			desc: "identical",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			want: ``,
-		},
-		{
-			desc: "first item changed",
-			a: `apiVersion: autoscaling/v2beta1
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			want: `apiVersion: autoscaling/v2beta1 -> autoscaling/v2
-`,
-		},
-		{
-			desc: "nested item changed",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-egressgateway
-    release: istio`,
-			want: `metadata:
-  labels:
-    app: istio-ingressgateway -> istio-egressgateway
-`,
-		},
-		{
-			desc: "one map value changed, order changed",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio
-spec:
-  maxReplicas: 5
-  minReplicas: 1
-  scaleTargetRef:
-    apiVersion: apps/v1
-    kind: Deployment
-    name: ingressgateway
-  metrics:
-    - type: Resource
-      resource:
-        name: cpu
-        target:
-          type: Utilization
-          averageUtilization: 80`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  labels:
-    app: istio-ingressgateway
-    release: istio
-  name: istio-ingressgateway
-  namespace: istio-system
-spec:
-  maxReplicas: 5
-  metrics:
-  - resource:
-      name: cpu
-      target:
-        type: Utilization
-        averageUtilization: 80
-    type: Resource
-  minReplicas: 1
-  scaleTargetRef:
-    apiVersion: apps/v1
-    kind: Deployment
-    name: istio-ingressgateway`,
-			want: `spec:
-  scaleTargetRef:
-    name: ingressgateway -> istio-ingressgateway
-`,
-		},
-		{
-			desc: "arrays with same items",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label2
-   - label3
-`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label2
-   - label3
-`,
-			want: ``,
-		},
-		{
-			desc: "arrays with different items",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label2
-   - label3
-`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label5
-   - label6
-`,
-			want: `metadata:
-  labels:
-    '[#1]': label2 -> label5
-    '[#2]': label3 -> label6
-`,
-		},
-		{
-			desc: "arrays with same items, order changed",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label2
-   - label3
-`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label2
-   - label3
-   - label1
-`,
-			want: `metadata:
-  labels:
-    '[?->2]': <empty> -> label1 (ADDED)
-    '[0->?]': label1 -> <empty> (REMOVED)
-`,
-		},
-		{
-			desc: "arrays with items",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    - label0
-    - label1
-    - label2
-`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    - label4
-    - label5
-    - label2
-    - label3
-    - label1
-    - label0
-`,
-			want: `metadata:
-  labels:
-    '[#0]': label0 -> label4
-    '[?->1]': <empty> -> label5 (ADDED)
-    '[?->2]': <empty> -> label2 (ADDED)
-    '[?->3]': <empty> -> label3 (ADDED)
-    '[2->5]': label2 -> label0
-`,
-		},
-		{
-			desc: "arrays with additional items",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label2
-   - label3
-`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label2
-   - label3
-   - label4
-   - label5
-`,
-			want: `metadata:
-  labels:
-    '[?->3]': <empty> -> label4 (ADDED)
-    '[?->4]': <empty> -> label5 (ADDED)
-`,
-		},
-		{
-			desc: "arrays with missing items",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label2
-   - label3
-   - label4
-   - label5
-`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
- name: istio-ingressgateway
- namespace: istio-system
- labels:
-   - label1
-   - label2
-   - label3
-`,
-			want: `metadata:
-  labels:
-    '[3->?]': label4 -> <empty> (REMOVED)
-    '[4->?]': label5 -> <empty> (REMOVED)
-`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			if got, want := YAMLCmp(tt.a, tt.b), tt.want; !(got == want) {
-				t.Errorf("%s: got:%v, want:%v", tt.desc, got, want)
-			}
-		})
-	}
-}
-
-func TestYAMLCmpWithIgnore(t *testing.T) {
-	tests := []struct {
-		desc string
-		a    string
-		b    string
-		i    []string
-		want any
-	}{
-		{
-			desc: "identical",
-			a: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			i:    []string{"metadata.annotations.checksum/config-volume"},
-			want: ``,
-		},
-		{
-			desc: "ignore checksum",
-			a: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 43d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 03ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			i:    []string{"metadata.annotations.checksum/config-volume"},
-			want: ``,
-		},
-		{
-			desc: "ignore missing checksum value",
-			a: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 43d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			i:    []string{"metadata.annotations.checksum/config-volume"},
-			want: ``,
-		},
-		{
-			desc: "ignore additional checksum value",
-			a: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 43d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			i:    []string{"metadata.annotations.checksum/config-volume"},
-			want: ``,
-		},
-		{
-			desc: "show checksum not exist",
-			a: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 43d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			i: []string{"metadata.annotations.checksum/config-volume"},
-			want: `metadata:
-  annotations: map[checksum/config-volume:43d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60]
-    -> <empty> (REMOVED)
-`,
-		},
-		{
-			desc: "ignore by wildcard",
-			a: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 01d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    checksum/config-volume: 02ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 03ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    checksum/config-volume: 04ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-    app: istio-ingressgateway
-    release: istio`,
-			i:    []string{"*.checksum/config-volume"},
-			want: ``,
-		},
-		{
-			desc: "ignore by wildcard negative",
-			a: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 01d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    checksum/config-volume: 02ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-    app: istio-ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 03ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    checksum/config-volume: 04ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-    app: istio-ingressgateway
-    release: istio`,
-			i: []string{"*labels.checksum/config-volume"},
-			want: `metadata:
-  annotations:
-    checksum/config-volume: 01d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-      -> 03ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-`,
-		},
-		{
-			desc: "ignore multiple paths",
-			a: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 43d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 03ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			i: []string{
-				"metadata.annotations.checksum/config-volume",
-				"metadata.labels.app",
-			},
-			want: ``,
-		},
-		{
-			desc: "ignore multiple paths negative",
-			a: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 43d72e930ed33e3e01731f8bcbf31dbf02cb1c1fc53bcc09199ab45c0d031b60
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: ingressgateway
-    release: istio`,
-			b: `apiVersion: autoscaling/v2
-kind: Deployment
-metadata:
-  annotations:
-    checksum/config-volume: 03ba6246b2c39b48a4f8c3a92c3420a0416804d38ebe292e65cf674fb0875192
-  name: istio-ingressgateway
-  namespace: istio-system
-  labels:
-    app: istio-ingressgateway
-    release: istio`,
-			i: []string{
-				"metadata.annotations.checksum/config-volume",
-			},
-			want: `metadata:
-  labels:
-    app: ingressgateway -> istio-ingressgateway
-`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			if got, want := YAMLCmpWithIgnore(tt.a, tt.b, tt.i, ""), tt.want; !(got == want) {
-				t.Errorf("%s: got:%v, want:%v", tt.desc, got, want)
-			}
-		})
-	}
-}
-
-func TestYAMLCmpWithIgnoreTree(t *testing.T) {
-	tests := []struct {
-		desc string
-		a    string
-		b    string
-		mask string
-		want any
-	}{
-		{
-			desc: "ignore masked",
-			a: `
-ak: av
-bk:
-  b1k: b1v
-  b2k: b2v
-`,
-			b: `
-ak: av
-bk:
-  b1k: b1v-changed
-  b2k: b2v-changed
-`,
-			mask: `
-bk:
-  b1k: ignored
-  b2k: ignored
-`,
-			want: ``,
-		},
-		{
-			desc: "ignore nested masked",
-			a: `
-ak: av
-bk:
-  bbk:
-    b1k: b1v
-    b2k: b2v
-`,
-			b: `
-ak: av
-bk:
-  bbk:
-    b1k: b1v-changed
-    b2k: b2v-changed
-`,
-			mask: `
-bk:
-  bbk:
-    b1k: ignored
-`,
-			want: `bk:
-  bbk:
-    b2k: b2v -> b2v-changed
-`,
-		},
-		{
-			desc: "not ignore non-masked",
-			a: `
-ak: av
-bk:
-  bbk:
-    b1k: b1v
-    b2k: b2v
-`,
-			b: `
-ak: av
-bk:
-  bbk:
-    b1k: b1v-changed
-    b2k: b2v
-`,
-			mask: `
-bk:
-  bbk:
-    b1k:
-      bbb1k: ignored
-`,
-			want: `bk:
-  bbk:
-    b1k: b1v -> b1v-changed
-`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			if got, want := YAMLCmpWithIgnore(tt.a, tt.b, nil, tt.mask), tt.want; !(got == want) {
-				t.Errorf("%s: got:%v, want:%v", tt.desc, got, want)
-			}
-		})
-	}
-}
-
-func TestYAMLCmpWithYamlInline(t *testing.T) {
-	tests := []struct {
-		desc string
-		a    string
-		b    string
-		want any
-	}{
-		{
-			desc: "ConfigMap data order changed",
-			a: `kind: ConfigMap
-data:
-  validatingwebhookconfiguration.yaml: |-
-    kind: ValidatingWebhookConfiguration
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    metadata:
-      name: istio-galley
-      value: foo`,
-			b: `kind: ConfigMap
-data:
-  validatingwebhookconfiguration.yaml: |-
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    kind: ValidatingWebhookConfiguration
-    metadata:
-      value: foo
-      name: istio-galley`,
-			want: ``,
-		},
-		{
-			desc: "ConfigMap data value change",
-			a: `kind: ConfigMap
-data:
-  validatingwebhookconfiguration.yaml: |-
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    kind: ValidatingWebhookConfiguration
-    metadata:
-      name: istio-galley`,
-			b: `kind: ConfigMap
-data:
-  validatingwebhookconfiguration.yaml: |-
-    kind: ValidatingWebhookConfiguration
-    apiVersion: admissionregistration.k8s.io/v1beta2
-    metadata:
-      name: istio-galley`,
-			want: `data:
-  validatingwebhookconfiguration.yaml:
-    apiVersion: admissionregistration.k8s.io/v1beta1 -> admissionregistration.k8s.io/v1beta2
-`,
-		},
-		{
-			desc: "ConfigMap data deep nested value change",
-			a: `apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: injector-mesh
-data:
-  mesh: |-
-    defaultConfig:
-      tracing:
-        zipkin:
-          address: zipkin.istio-system:9411
-      controlPlaneAuthPolicy: NONE
-      connectTimeout: 10s`,
-			b: `apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: injector-mesh
-data:
-  mesh: |-
-    defaultConfig:
-      connectTimeout: 10s
-      tracing:
-        zipkin:
-          address: zipkin.istio-system:9412
-      controlPlaneAuthPolicy: NONE`,
-			want: `data:
-  mesh:
-    defaultConfig:
-      tracing:
-        zipkin:
-          address: zipkin.istio-system:9411 -> zipkin.istio-system:9412
-`,
-		},
-		{
-			desc: "ConfigMap data multiple changes",
-			a: `apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: injector-mesh
-  namespace: istio-system
-  labels:
-    release: istio
-data:
-  mesh: |-
-    defaultConfig:
-      connectTimeout: 10s
-      configPath: "/etc/istio/proxyv1"
-      serviceCluster: istio-proxy
-      drainDuration: 45s
-      proxyAdminPortA: 15000
-      concurrency: 2
-      tracing:
-        zipkin:
-          address: zipkin.istio-system:9411
-      controlPlaneAuthPolicy: NONE
-      discoveryAddress: istio-pilot.istio-system:15010`,
-			b: `apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: injector-mesh
-  namespace: istio-system
-  labels:
-    release: istio
-data:
-  mesh: |-
-    defaultConfig:
-      connectTimeout: 10s
-      configPath: "/etc/istio/proxyv2"
-      serviceCluster: istio-proxy
-      drainDuration: 45s
-      proxyAdminPortB: 15000
-      concurrency: 2
-      tracing:
-        zipkin:
-          address: zipkin.istio-system:9411
-      controlPlaneAuthPolicy: NONE
-      discoveryAddress: istio-pilot.istio-system:15010`,
-			want: `data:
-  mesh:
-    defaultConfig:
-      configPath: /etc/istio/proxyv1 -> /etc/istio/proxyv2
-      proxyAdminPortA: 15000 -> <empty> (REMOVED)
-      proxyAdminPortB: <empty> -> 15000 (ADDED)
-`,
-		},
-		{
-			desc: "Not ConfigMap, identical",
-			a: `kind: Config
-data:
-  validatingwebhookconfiguration.yaml: |-
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    kind: ValidatingWebhookConfiguration
-    metadata:
-      name: istio-galley`,
-			b: `kind: Config
-data:
-  validatingwebhookconfiguration.yaml: |-
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    kind: ValidatingWebhookConfiguration
-    metadata:
-      name: istio-galley`,
-			want: ``,
-		},
-		{
-			desc: "Not ConfigMap, Order changed",
-			a: `kind: Config
-data:
-  validatingwebhookconfiguration.yaml: |-
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    kind: ValidatingWebhookConfiguration
-    metadata:
-      name: istio-galley`,
-			b: `kind: Config
-data:
-  validatingwebhookconfiguration.yaml: |-
-    kind: ValidatingWebhookConfiguration
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    metadata:
-      name: istio-galley`,
-			want: `data:
-  validatingwebhookconfiguration.yaml: |-
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    kind: ValidatingWebhookConfiguration
-    metadata:
-      name: istio-galley -> kind: ValidatingWebhookConfiguration
-    apiVersion: admissionregistration.k8s.io/v1beta1
-    metadata:
-      name: istio-galley
-`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			if got, want := YAMLCmp(tt.a, tt.b), tt.want; !(got == want) {
-				t.Errorf("%s: got:%v, want:%v", tt.desc, got, want)
-			}
-		})
-	}
-}
